Conflicts:
	Source/Core/VideoCommon/AVIDump.cpp

+++ Source/Core/Core/HLE/HLE_OS.cpp
@@ -2,14 +2,14 @@
 // Licensed under GPLv2
 // Refer to the license.txt file included.
 
-#include "StringUtil.h"
 #include <string>
 
-#include "Common.h"
-#include "HLE_OS.h"
+#include "Common/Common.h"
+#include "Common/StringUtil.h"
 
-#include "../PowerPC/PowerPC.h"
-#include "../HW/Memmap.h"
+#include "Core/HLE/HLE_OS.h"
+#include "Core/HW/Memmap.h"
+#include "Core/PowerPC/PowerPC.h"
 
 namespace HLE_OS
 {
@@ -32,9 +32,12 @@ void HLE_OSPanic()
 void HLE_GeneralDebugPrint()
 {
 	std::string ReportMessage;
-	if(*(u32*)Memory::GetPointer(GPR(3)) > 0x80000000){
+	if (*(u32*)Memory::GetPointer(GPR(3)) > 0x80000000)
+	{
 		GetStringVA(ReportMessage, 4);
-	}else{
+	}
+	else
+	{
 		GetStringVA(ReportMessage);
 	}
 	NPC = LR;
@@ -50,7 +53,8 @@ void HLE_VPrintf()
 	u32 offset = Memory::Read_U32(r4+8);
 	u32 check = Memory::Read_U32(r4);
 	//NOTICE_LOG(OSREPORT, "Offset: %08X, Check %08X", offset, check);
-	for(int i = 4; i<= 10; i++){
+	for (int i = 4; i<= 10; i++)
+	{
 		GPR(i) = Memory::Read_U32(offset+(i-(check == 0x01000000? 3 : 2))*4);
 		//NOTICE_LOG(OSREPORT, "r%d: %08X",i, GPR(i));
 	}
@@ -96,18 +100,18 @@ void GetStringVA(std::string& _rOutBuffer, u32 strReg)
 		return;
 	}
 
-	while(*pString)
+	while (*pString)
 	{
 		if (*pString == '%')
 		{
 			char* pArgument = ArgumentBuffer;
 			*pArgument++ = *pString++;
-			if(*pString == '%') {
+			if (*pString == '%') {
 				_rOutBuffer += "%";
 				pString++;
 				continue;
 			}
-			while(*pString < 'A' || *pString > 'z' || *pString == 'l' || *pString == '-')
+			while (*pString < 'A' || *pString > 'z' || *pString == 'l' || *pString == '-')
 				*pArgument++ = *pString++;
 
 			*pArgument++ = *pString;
@@ -130,7 +134,7 @@ void GetStringVA(std::string& _rOutBuffer, u32 strReg)
 			}
 			ParameterCounter++;
 
-			switch(*pString)
+			switch (*pString)
 			{
 			case 's':
 				_rOutBuffer += StringFromFormat(ArgumentBuffer, (char*)Memory::GetPointer((u32)Parameter));
@@ -155,7 +159,7 @@ void GetStringVA(std::string& _rOutBuffer, u32 strReg)
 
 			case 'p':
 				// Override, so 64bit dolphin prints 32bit pointers, since the ppc is 32bit :)
-				_rOutBuffer += StringFromFormat("%x", Parameter);
+				_rOutBuffer += StringFromFormat("%x", (u32)Parameter);
 				break;
 
 			default:
@@ -170,7 +174,7 @@ void GetStringVA(std::string& _rOutBuffer, u32 strReg)
 			pString++;
 		}
 	}
-	if(_rOutBuffer[_rOutBuffer.length() - 1] == '\n')
+	if (_rOutBuffer[_rOutBuffer.length() - 1] == '\n')
 		_rOutBuffer.resize(_rOutBuffer.length() - 1);
 }
